move time label from header to reload button
